+++ packages/channels/botpress-channel-web/src/views/web/index.jsx
@@ -8,6 +8,8 @@ import classnames from 'classnames'
 import addMilliseconds from 'date-fns/add_milliseconds'
 import isBefore from 'date-fns/is_before'
 import queryString from 'query-string'
+import moment from 'moment'
+import ms from 'ms'
 
 import Convo from './convo'
 import Side from './side'
@@ -35,7 +37,10 @@ const defaultOptions = {
   foregroundColor: '#000000',
   textColorOnForeground: '#ffffff',
   enableReset: false,
-  botConvoTitle: 'Botpress Webchat'
+  showUserName: false,
+  showUserAvatar: false,
+  botConvoTitle: 'Botpress Webchat',
+  enableTranscriptDownload: false
 }
 
 export default class Web extends React.Component {
@@ -66,7 +71,7 @@ export default class Web extends React.Component {
 
   componentDidMount() {
     this.setUserId()
-      .then(::this.fetchData)
+      .then(this.fetchData)
       .then(() => {
         this.handleSwitchView('widget')
         if (!this.state.isButtonHidden) {
@@ -201,7 +206,7 @@ export default class Web extends React.Component {
     }, ANIM_DURATION * 2.1)
   }
 
-  handleButtonClicked() {
+  handleButtonClicked = () => {
     if (this.state.view === 'convo') {
       this.handleSwitchView('widget')
     } else {
@@ -215,12 +220,19 @@ export default class Web extends React.Component {
       this.props.bp.events.setup()
     }
 
-    this.props.bp.events.on('guest.webchat.message', ::this.handleNewMessage)
-    this.props.bp.events.on('guest.webchat.typing', ::this.handleBotTyping)
+    this.props.bp.events.on('guest.webchat.message', this.handleNewMessage)
+    this.props.bp.events.on('guest.webchat.typing', this.handleBotTyping)
   }
 
-  fetchData() {
-    return this.fetchConversations().then(::this.fetchCurrentConversation)
+  fetchData = () => {
+    return this.fetchConversations()
+      .then(this.fetchCurrentConversation)
+      .then(() => {
+        this.handleSendData({
+          type: 'visit',
+          text: 'User visit'
+        })
+      })
   }
 
   fetchConversations() {
@@ -228,18 +240,21 @@ export default class Web extends React.Component {
     const userId = this.userId
     const url = `${BOT_HOSTNAME}/api/botpress-platform-webchat/conversations/${userId}`
 
-    return axios
-      .get(url)
-      .then(({ data }) => new Promise(resolve => !this.isUnmounted && this.setState({ conversations: data }, resolve)))
+    return axios.get(url).then(({ data }) => new Promise(resolve => !this.isUnmounted && this.setState(data, resolve)))
   }
 
-  fetchCurrentConversation(convoId) {
+  fetchCurrentConversation = convoId => {
     const axios = this.props.bp.axios
     const userId = this.userId
+    const { conversations, currentConversationId } = this.state
 
-    let conversationIdToFetch = convoId || this.state.currentConversationId
-    if (this.state.conversations.length > 0 && !conversationIdToFetch) {
-      conversationIdToFetch = this.state.conversations[0].id
+    let conversationIdToFetch = convoId || currentConversationId
+    if (conversations.length > 0 && !conversationIdToFetch) {
+      const lifeTimeMargin = moment().subtract(ms(this.state.recentConversationLifetime), 'ms')
+      if (moment(conversations[0].last_heard_on).isBefore(lifeTimeMargin) && this.state.startNewConvoOnTimeout) {
+        return
+      }
+      conversationIdToFetch = conversations[0].id
       this.setState({ currentConversationId: conversationIdToFetch })
     }
 
@@ -256,7 +271,11 @@ export default class Web extends React.Component {
     })
   }
 
-  handleNewMessage(event) {
+  handleNewMessage = event => {
+    if (event.message_type === 'visit') {
+      // don't do anything, it's the system message
+      return
+    }
     this.safeUpdateCurrentConvo(event.conversationId, true, convo => {
       return Object.assign({}, convo, {
         messages: [...convo.messages, event],
@@ -265,17 +284,17 @@ export default class Web extends React.Component {
     })
   }
 
-  handleBotTyping(event) {
+  handleBotTyping = event => {
     this.safeUpdateCurrentConvo(event.conversationId, false, convo => {
       return Object.assign({}, convo, {
         typingUntil: addMilliseconds(new Date(), event.timeInMs)
       })
     })
 
-    setTimeout(::this.expireTyping, event.timeInMs + 50)
+    setTimeout(this.expireTyping, event.timeInMs + 50)
   }
 
-  expireTyping() {
+  expireTyping = () => {
     const currentTypingUntil = this.state.currentConversation && this.state.currentConversation.typingUntil
 
     const timerExpired = currentTypingUntil && isBefore(new Date(currentTypingUntil), new Date())
@@ -289,7 +308,7 @@ export default class Web extends React.Component {
   safeUpdateCurrentConvo(convoId, addToUnread, updater) {
     // there's no conversation to update or our convo changed
     if (!this.state.currentConversation || this.state.currentConversationId !== convoId) {
-      this.fetchConversations().then(::this.fetchCurrentConversation)
+      this.fetchConversations().then(this.fetchCurrentConversation)
 
       return
     }
@@ -336,7 +355,7 @@ export default class Web extends React.Component {
     })
   }
 
-  handleResetUnreadCount() {
+  handleResetUnreadCount = () => {
     if (document.hasFocus && document.hasFocus() && this.state.view === 'side') {
       this.setState({
         unreadCount: 0
@@ -344,29 +363,26 @@ export default class Web extends React.Component {
     }
   }
 
-  handleSendMessage() {
-    const userId = window.__BP_VISITOR_ID
-    const config = { params: { conversationId: this.state.currentConversationId } }
-
+  handleSendMessage = () => {
     return this.handleSendData({ type: 'text', text: this.state.textToSend }).then(() => {
       this.handleSwitchView('side')
       this.setState({ textToSend: '' })
     })
   }
 
-  handleTextChanged(event) {
+  handleTextChanged = event => {
     this.setState({
       textToSend: event.target.value
     })
   }
 
-  handleAddEmoji(emoji, event) {
+  handleAddEmoji = emoji => {
     this.setState({
       textToSend: this.state.textToSend + emoji.native + ' '
     })
   }
 
-  handleSendQuickReply(title, payload) {
+  handleSendQuickReply = (title, payload) => {
     return this.handleSendData({
       type: 'quick_reply',
       text: title,
@@ -374,7 +390,7 @@ export default class Web extends React.Component {
     })
   }
 
-  handleSendForm(fields, formId, repr) {
+  handleSendForm = (fields, formId, repr) => {
     return this.handleSendData({
       type: 'form',
       formId: formId,
@@ -383,7 +399,7 @@ export default class Web extends React.Component {
     })
   }
 
-  handleLoginPrompt(username, password) {
+  handleLoginPrompt = (username, password) => {
     return this.handleSendData({
       type: 'login_prompt',
       text: 'Provided login information',
@@ -391,7 +407,7 @@ export default class Web extends React.Component {
     })
   }
 
-  handleFileUploadSend(title, payload, file) {
+  handleFileUploadSend = (title, payload, file) => {
     const userId = window.__BP_VISITOR_ID
     const url = `${BOT_HOSTNAME}/api/botpress-platform-webchat/messages/${userId}/files`
     const config = { params: { conversationId: this.state.currentConversationId } }
@@ -409,7 +425,7 @@ export default class Web extends React.Component {
     return this.props.bp.axios.post(url, data, config).then()
   }
 
-  handleSwitchConvo(convoId) {
+  handleSwitchConvo = convoId => {
     this.setState({
       currentConversation: null,
       currentConversationId: convoId
@@ -418,11 +434,11 @@ export default class Web extends React.Component {
     this.fetchCurrentConversation(convoId)
   }
 
-  handleClosePanel() {
+  handleClosePanel = () => {
     this.handleSwitchView('widget')
   }
 
-  handleSessionReset() {
+  handleSessionReset = () => {
     const userId = window.__BP_VISITOR_ID
     const url = `${BOT_HOSTNAME}/api/botpress-platform-webchat/conversations/${userId}/${this.state
       .currentConversationId}/reset`
@@ -464,7 +480,7 @@ export default class Web extends React.Component {
     return (
       <button
         className={style[this.state.widgetTransition]}
-        onClick={::this.handleButtonClicked}
+        onClick={this.handleButtonClicked}
         style={{ backgroundColor: this.state.config.foregroundColor }}
       >
         <i>{this.state.view === 'convo' ? this.renderCloseIcon() : this.renderOpenIcon()}</i>
@@ -481,8 +497,8 @@ export default class Web extends React.Component {
             {this.state.view === 'convo' ? (
               <Convo
                 transition={this.state.convoTransition}
-                change={::this.handleTextChanged}
-                send={::this.handleSendMessage}
+                change={this.handleTextChanged}
+                send={this.handleSendMessage}
                 config={this.state.config}
                 text={this.state.textToSend}
               />
@@ -492,6 +508,30 @@ export default class Web extends React.Component {
         </div>
       </div>
     )
+  }
+
+  downaloadFile(name, blob) {
+    const url = window.URL.createObjectURL(blob)
+    const link = document.createElement('a')
+
+    link.href = url
+    link.setAttribute('download', name)
+
+    document.body.appendChild(link)
+    link.click()
+
+    document.body.removeChild(link)
+    window.URL.revokeObjectURL(url)
+  }
+
+  downloadConversation = async () => {
+    const userId = window.__BP_VISITOR_ID
+    const url = `${BOT_HOSTNAME}/api/botpress-platform-webchat/conversations/${userId}/${this.state
+      .currentConversationId}/download/txt`
+    const file = (await this.props.bp.axios.get(url)).data
+    const blobFile = new Blob([file.txt])
+
+    this.downaloadFile(file.name, blobFile)
   }
 
   renderSide() {
@@ -504,17 +544,18 @@ export default class Web extends React.Component {
         unreadCount={this.state.unreadCount}
         currentConversation={this.state.currentConversation}
         conversations={this.state.conversations}
-        addEmojiToText={::this.handleAddEmoji}
-        onClose={!this.props.fullscreen ? ::this.handleClosePanel : null}
-        onResetSession={::this.handleSessionReset}
-        onSwitchConvo={::this.handleSwitchConvo}
-        onTextSend={::this.handleSendMessage}
-        onTextChanged={::this.handleTextChanged}
-        onQuickReplySend={::this.handleSendQuickReply}
-        onFormSend={::this.handleSendForm}
-        onFileUploadSend={::this.handleFileUploadSend}
-        onLoginPromptSend={::this.handleLoginPrompt}
-        onSendData={::this.handleSendData}
+        addEmojiToText={this.handleAddEmoji}
+        onClose={!this.props.fullscreen ? this.handleClosePanel : null}
+        onResetSession={this.handleSessionReset}
+        onSwitchConvo={this.handleSwitchConvo}
+        onTextSend={this.handleSendMessage}
+        onTextChanged={this.handleTextChanged}
+        onQuickReplySend={this.handleSendQuickReply}
+        onFormSend={this.handleSendForm}
+        onFileUploadSend={this.handleFileUploadSend}
+        onLoginPromptSend={this.handleLoginPrompt}
+        onSendData={this.handleSendData}
+        downloadConversation={this.downloadConversation}
       />
     )
   }
@@ -530,7 +571,7 @@ export default class Web extends React.Component {
     const view = this.state.view !== 'side' && !this.props.fullscreen ? this.renderWidget() : this.renderSide()
 
     return (
-      <div className={style.web} onFocus={::this.handleResetUnreadCount}>
+      <div className={style.web} onFocus={this.handleResetUnreadCount}>
         {view}
       </div>
     )