+++ packages/core/botpress/src/botpress.js
@@ -68,9 +68,7 @@ import packageJson from '../package.json'
 
 import createServer from './server'
 
-import { getDataLocation, getBotpressVersion } from './util'
-
-import { isDeveloping, print } from './util'
+import { getDataLocation, getBotpressVersion, validateBotVersion, isDeveloping, print } from './util'
 
 const RESTART_EXIT_CODE = 107
 
@@ -121,6 +119,8 @@ const validateBotfile = botfile => {
     )
     throw new Error('Outdated botfile format')
   }
+
+  validateBotVersion(packageJson.version, botfile.version)
 
   for (const prop of REQUIRED_PROPS) {
     if (!(prop in botfile)) {