+++ packages/functionals/botpress-hitl/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="10.28.0"></a>
+# [10.28.0](https://github.com/botpress/modules/compare/v10.27.1...v10.28.0) (2018-08-03)
+
+
+### Bug Fixes
+
+* **hitl:** sqlite alert ([544aa41](https://github.com/botpress/modules/commit/544aa41))
+
+
+
+
 <a name="10.27.1"></a>
 ## [10.27.1](https://github.com/botpress/modules/compare/v10.27.0...v10.27.1) (2018-08-01)